requirements.txt
# 核心依赖
a2a>=0.44
agno>=1.7.6
fastapi>=0.116.0
openai>=1.93.3
a2a-sdk>=0.2.16
uvicorn>=0.35.0
httpx>=0.24.0
pydantic>=2.0.0
alembic>=1.12.0
sqlalchemy>=2.0.0
python-dotenv==1.0.1